# Port ARM inflate performance improvement patches (chunk SIMD, read64le)