… inffast_chunk.c
… inflate.h
… zlib.h