This series ports ARM inflate performance improvement patches. With this, the performance improvement during inflate, tested using a modified zpipe.c and the Silesia corpus, is around 17-34%.

Patches 1-3 include a bug-fix and some code improvements taken from madler/zlib.

Patch 4 is a code readability port from zlib chromium.

    * Chris Blume (cblume@chromium):
      8888511 - "Zlib: Use defines for inffast"
      b9c1566 - "Share inffast names in zlib"
    
Patch 5 introduces 3 new files: inffast_chunk.c, inffast_chunk.h and chunkcopy.h.
These incorporate the changes from 2 patches in zlib chromium.
    
    * Adenilson Cavalcanti (adenilson.cavalcanti@arm.com)
      3060dcb - "zlib: inflate using wider loads and stores"
    
    * Noel Gordon (noel@chromium.org)
      64ffef0 - "Improve zlib inflate speed by using SSE2 chunk copy"

Patch 6 is a port of a performance improvement patch from zlib chromium. 

    * Noel Gordon (noel@chromium.org)
      8a8edc1 - "Increase inflate speed: read decoder input into a uint64_t"