@janaknat 2020-08-26
Some inflate performance number breakdown (average of 5 runs):

Dataset| File | Throughput before (kb/ms)| Throughput after (kb/ms)| % change |
--------|-----|--------------------|----------------|---|
Silesia | xml | 70.86 | 94.81| +34%
 | | x-ray | 138.68 | 155.01| +12%
 | | webster | 83.47 | 108.71| +30%
 | | sao | 154.70 | 180.67 | +17%
 | | samba | 94.49 | 122.20 | +30%
 | | reymont | 82.65 | 108.34 | +31%
 | | osdb | 108.96 | 128.85 | +18%
 | | ooffice | 107.34 | 126.89 | +17%
 | | nci | 61.99 | 82.12 | +34%
 | | mr | 94.72 | 122.25 | +29%
 | | mozilla | 107.81 | 131.39 | +22%
 | | dickens | 88.84 | 118.96 | +34%