@vkrasnov 2020-09-04
Nice, what CPU was used for the benchmarks? It looks like Intel should be faster as well?