@vkrasnov 2020-09-23
This is great. I actually managed to apply some of the optimizations to Intel too for a 16% inflate speedup. I will apply those too.