@janaknat 2020-09-23
@vkrasnov Thanks. I had the Intel optimizations lined up next. Was waiting for this PR to land. If you've got that covered, that's great.