@vkrasnov 2020-09-23
I am happy to accept another PR, I just have a PoC