@janaknat 2020-09-24
@vkrasnov I've created a PR with the Intel optimizations.